Separate the idea of a packet from the idea of a payload
A packet contains a payload. Once we've extracted the packet from the
stream, and verified its CRC, we need to decode the payload. The
payload doesn't know about the packer header and CRC, nor does it need
to. By the time we have the payload, those have already been checked.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>