sickd: Implement infrastructure for event-driven loops
sickd is designed to be light on both memory and CPU usage. To achieve
the low CPU usage requirement, sickd should sleep waiting for data to
arrive. This is achieved with the poll() system call.

We need a way to add and remove event sources, and call each device's
process_events() function when something happens. Internally, this is
implented with a map-vector hybrid data structure., the details of
which are explained in the comments

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>